pyproject.toml
[project]
name = "ag-worldbank-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.36",
    "chromadb>=1.1.0",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-groq>=0.3.8",
    "pymupdf>=1.26.4",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "temp>=2020.7.2",
]
